+++ src/lib/preferences/activity.ts
@@ -15,10 +15,15 @@ export type Verification = {
 	properties: string[];
 };
 
+export type ExpiredCredential = {
+    type: 'expired';
+    id: number;
+};
+
 export type Activity = {
 	at: number;
 	read?: boolean;
-} & (IssuedCredential | Verification);
+} & (IssuedCredential | Verification | ExpiredCredential);
 
 export const ACTIVITY_PREFERENCES_KEY = 'activity';