+++ src/lib/tabs/IonTabs.svelte
@@ -4,8 +4,8 @@
 	import { onMount } from 'svelte';
 	import { page, navigating } from '$app/stores';
 
-	import { goto } from '$app/navigation';
-
+	import { goto, i18n, r } from '$lib/i18n';
+	import { goto as svelteGoto } from '$app/navigation';
 	export let ionTabsDidChange = () => {};
 	export let ionTabsWillChange = () => {};
 
@@ -22,14 +22,14 @@
 	const { pathname } = $page.url;
 	const pathSplit = pathname.split('/');
 	let currentTabName = pathSplit[pathSplit.length - 1]; // we don't want to use at(-1) because of old browsers
-	const relativePath = pathname.replace(currentTabName, '');
+	const relativePath = i18n.route(pathname.replace(currentTabName, ''));
 
 	// we need to capture the router changes - to support a-href navigation and other stuff
 	$: if ($navigating && $navigating.to) {
 		tabs.forEach(async (tab) => {
 			if ($navigating.to.url.pathname.includes(relativePath + tab.tab)) {
 				currentTabName = tab.tab;
-				await goto(relativePath + tab.tab);
+				await svelteGoto(r(relativePath + tab.tab));
 				controller.select(tab.tab);
 			}
 		});